Improve enum class handling.

+++ src/PBR.h
@@ -80,7 +80,7 @@ typedef std::int16_t int16;
 typedef std::int32_t int32;
 typedef std::int64_t int64;
 
-typedef int64 RRID;
+typedef uint32 RRID;
 
 template<typename T>
 using sref = std::shared_ptr<T>;
